Rename core to kueue
To play well with kubebuilder
