+++ pkg/queue/cluster_queue_best_effort_fifo.go
@@ -19,7 +19,7 @@ package queue
 import (
 	"k8s.io/apimachinery/pkg/api/equality"
 
-	kueue "sigs.k8s.io/kueue/apis/core/v1alpha1"
+	kueue "sigs.k8s.io/kueue/apis/kueue/v1alpha1"
 	"sigs.k8s.io/kueue/pkg/workload"
 )
