+++ pkg/queue/queue.go
@@ -19,7 +19,7 @@ package queue
 import (
 	"fmt"
 
-	kueue "sigs.k8s.io/kueue/apis/core/v1alpha1"
+	kueue "sigs.k8s.io/kueue/apis/kueue/v1alpha1"
 	"sigs.k8s.io/kueue/pkg/workload"
 )
