+++ pkg/workload/workload_test.go
@@ -29,7 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	utiltesting "sigs.k8s.io/kueue/pkg/util/testing"
 
-	kueue "sigs.k8s.io/kueue/apis/core/v1alpha1"
+	kueue "sigs.k8s.io/kueue/apis/kueue/v1alpha1"
 )
 
 func TestPodRequests(t *testing.T) {
